Refactor add bounty modalize
